fix(synthetic-shadow): move activeElement test to karma
