+++ packages/integration-karma/test/polyfills/shadow-root/x/test/test.html
@@ -1,0 +1,3 @@
+<template>
+    <input />
+</template>
