+++ packages/integration-karma/test/polyfills/shadow-root/x/test/test.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class Test extends LightningElement {}
